requirements.txt
# For Apple Silicon (M1/M2/M3)
tensorflow-macos==2.16.1
tensorflow-metal==1.1.0
protobuf<4.0.0

# For CUDA systems (NVIDIA GPUs)
# tensorflow[and-cuda]==2.16.1

# Core dependencies
numpy==1.26.3
pillow==10.2.0
click==8.1.7

# Web application
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.3
